Cake model specifies the vertical velocity v_z
Originally, in_params%vmacro specifies the LoS velocity in the constant
property slab.  In the non-linear synthesis mode, the LoS velocity will
be calculated from the vertical velocity v_z, specified in each slab of
the cake model, while the v_z velocity will be used in the iterative
solution of the transfer equation.

Rename %velocity to %v_z in VARS/type_slab, in IO/read_slab(), and in
TRANSFER/do_transfer().