Fix mu angle in the final formal solution
The final formal solution, which integrates the transfer equation for
the observed Stokes profiles, must follow the (theta, chi, gamma)
direction.  In fact, the RT_COEF/calc_rt_coef() computes the
emissivities and absorptivities for this direction, but the current
formal solution is performed for theta = 0.

In TRANSFER/synthesize_stokes() set mu to $\cos\theta$ so that the
optical depths, computed from slab%z heights, increase for an inclined
ray and the boundary condition from I0_allen() is set for the right
angle.