Change the angle quadrature
It turns out that the problem has to be solved in 3D geometry as the
radiative transfer quantities are dependent on the azimuth.  We have to
employ a proper angle quadruature set of directions, densely covering
the whole $4\pi$ sphere.  At the moment we will do tests using the
Lebedev quadratures.  Initially, they will be provided as a table of
$(\chi, \theta, w)$ values in the slab.model file.  Later we will set
them inside the code.

In VARS, remove %nmus_photosphere, %nmus_nophotosphere, and %mus(:)
fileds from the type_slab type and add instead %aq_size, %aq_azimuth(:),
%aq_inclination(:), and %aq_weight(:).

In IO/read_slab(), change the order of reading for the table of slab
parameters and the table of the angle quadrature.

At the moment, the rest of the code in TRANSFER/do_transfer() is not
operating correctly.