Add spectral properties to slab.mod
The non-linear transfer must be performed in each multiplet of the model
atom.  For each of them one must specify the multiplet wavelength, the
wavelength interval and the number of grid points for the wavelength
quadrature.  Add this funtionality to slab.mod and to IO/read_slab().

The corresponding structure must also contain indices (start/end) of the
bounding frequencies in the whole spectrum.