Change reduced frequency evaluation in calc_rt_coef()
The current version of the code in RT_COEF/calc_rt_coef() uses a legacy
trick which came from the original code by Egidio.  The reduced
frequency, which appears in the calculation of the complex absorption
profile,

  prof = profile(adamp,(onum0-onum-va)/dnum) / (dnum_freq*SQRTPI)

in three quantities, onum0, onum, and va, has units of linear
wavenumbers usually denoted in the literature by $\tilde{\nu}$ or
$\sigma$ and which is a reciprocal of the wavelength.  The corresponding
formulae for computing the wavelengths %wl, the frequencies %freq, and
the wavenumbers onum, are confusing because the look different from the
standard formulae when usual frequency or wavelength units are involved.

The legacy comes from the paschen() subroutine that outputs two arrays
autl and autu.  These arrays provide energies of the states, split in
the magnetic field, expressed in cm^{-1}.  The difference of these
energies between the upper and the lower state goes to onum0,

  onum0 = autu(mu2,jsmallu) - autl(ml2,jsmalll)

which is also in cm^{-1}.

To avoid future confusions with these physical units, let's express all
these quantities as reduced frequences,
\[
  v = \dfrac{ \nu }{ \Delta\nu_D }.
\]

First, rename %omin and %omax to %d_wl_min and %d_wl_max as they must
contain wavelength shifts off the nominal line center and not the
wavenumbers.

Second, generate the wavelength and frequency quadratures in the
beginning of TRANSFER/do_transfer() for each multiplet and store them in
%d_lambdas, %d_nus fields added to the multiplet type in VARS.

Third, replace onum, onum0, and va with the corresponding quantities
called red_freqs, red_freq_0, and red_freq_a, keeping reduced
frequencies instead of confusing wavenumbers.  The reduced frequency is
a dimensionless quantity given by the linear frequency divided by the
Doppler width in the same frequency units.