Set the outer loops for the fill_SEE()
Put a sequence of several nested loops in the right order so that to
minimize the number of calls to fill_SEE().  Keep in mind that the
system of statistical equilibrium equations is set only for one current
layer so you have to iterate.  The same is true for calc_rt_coef(),
which gives the RT data only for the specified direction.

Rename relative_change to tolerance, loop_iteration to iteration.  Add
extra temporary storage variables called save_* with the names of the
fields that needs to be preserved between the iterations.