Reallocate arrays in calc_rt_coef()
As we compute the RT output for many different lines, the output
arrays computed inside calc_rt_coef() must be reallocated for
different sizes of the frequency grid.  At the moment the code
uses the same arrays allocated for the first multiplet (10827).

Set all arrays to be completely deallocated and re-allocated in the
beginning of the calc_rt_coef().  This is time-inefficient so think
in the future how to use static arrays instead.