Fix %use_atomic_pol key
There are two possible values of the %use_atomic_pol key: 0 means
neglect all effects of atomic polarization and 1 means include them.

There are code pieces showing that the key might have a value of -1
to neglect the anisotropy of the radiation field in the SEEs.  At the
same time, there are if-then branches showing that the expected value
might be 2.

Fix the code to have the value of 1 by default to include the effects
of atomic level polarization, the other value of 0 to switch these
effects off, and one optional value of -1 to neglect the anisotropy
in the SEEs.