Add book-keeping for nbar and omega
The last sequence of nested loops, which has the formal solution, must
evaluate incrementally for each ray the corresponding radiation field
tensors Jbar00 and Jbar20, from which at the end of the last loop the
corresponding parameters nbar and omega should be evaluated for each
multiplet.  The maximal relative changes in these parameters, max
|dn/n| and max |dw/w|, must control the tolerance of the iterative
process of lambda-iterations.  Typically ~10 iterations are enough to
converge to the 0.1% relative precisions both in nbar and omega.

For this add several spectral arrays such as Ibar00 and Ibar20, which
keep the corresponding combinations of Stokes I and Stokes I,Q,U
augmented by the corresponding absorption profile along the ray, from
which the Jbar00 and Jbar20 tensors are evaluated.  At the first layer
the thensors are computed separately, outside of the loop along layers,
while in the other layers they are computed inside the inner formal solution
loop.

This is a preliminary version.  Many tests and optimisations will follow
soon.